use setupFiles
