+++ jest.config.js
@@ -8,6 +8,7 @@ module.exports = {
       useESM: true,
     },
   },
+  setupFiles: ['./test/setupFiles.ts'],
   moduleNameMapper: {
     "^(\\.{1,2}/.*)\\.js$": "$1",
     "\\.(jpg|ico|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|ogg)$":
